Uses shorts intead than integers for resource keys. This is sufficient and avoid to fill the constants pool in ".class" files.
git-svn-id: https://svn.apache.org/repos/asf/sis/branches/JDK7@1550874 13f79535-47bb-0310-9956-ffa450edef68

+++ core/sis-utility/src/main/java/org/apache/sis/math/MathFunctions.java
@@ -682,7 +682,7 @@ public final class MathFunctions extends Static {
         if (ordinal >= MIN_NAN_ORDINAL && ordinal <= MAX_NAN_ORDINAL) {
             return ordinal;
         }
-        final int resourceKey;
+        final short resourceKey;
         final Object obj;
         if (Float.isNaN(value)) {
             resourceKey = Errors.Keys.IllegalBitsPattern_1;
